+++ RELEASE-NOTES.md
@@ -3,6 +3,14 @@ different releases and which versions of PHP and MediaWiki they support, see the
 [platform compatibility tables](INSTALL.md#platform-compatibility-and-release-status).
 
 
+## maps 10.0.0
+
+Under development
+
+* Added support for MediaWiki 1.38.x
+* Packages installed directly into `Maps/vendor/` by executing Composer in `Maps/` are no longer loaded
+* Removed `Maps_VERSION` and `SM_VERSION` PHP constants
+
 ## Maps 9.0.7
 
 Released on February 18th, 2022.